requirements.txt
Django>=4.2
requests
python-dotenv
google-generativeai
